+++ src/main/java/io/github/xf8b/morefeatures/config/ServerConfig.java
@@ -4,11 +4,6 @@ import io.github.xf8b.morefeatures.MoreFeatures;
 import net.minecraftforge.common.ForgeConfigSpec;
 
 public class ServerConfig {
-    //Generation Settings
-    public final ForgeConfigSpec.BooleanValue isFeatureGenerationEnabled;
-    public final ForgeConfigSpec.BooleanValue isOreGenerationEnabled;
-    public final ForgeConfigSpec.BooleanValue isTreeGenerationEnabled;
-
     //Enchantment Settings
     public final ForgeConfigSpec.IntValue savingGraceChanceIncrease;
     public final ForgeConfigSpec.IntValue soulsRequiredForSharpnessLevelUp;
@@ -30,21 +25,6 @@ public class ServerConfig {
 
     public ServerConfig(ForgeConfigSpec.Builder builder) {
         builder.comment("The config file for MoreFeatures. Edit values below to change how certain things work");
-        builder.push("Generation Settings");
-        isFeatureGenerationEnabled = builder
-                .comment("Turns feature generation on or off")
-                .translation(MoreFeatures.MOD_ID + ".config." + "isFeatureGenerationEnabled")
-                .define("isFeatureGenerationEnabled", true);
-        isOreGenerationEnabled = builder
-                .comment("Turns ore generation on or off")
-                .translation(MoreFeatures.MOD_ID + ".config." + "isOreGenerationEnabled")
-                .define("isOreGenerationEnabled", true);
-        isTreeGenerationEnabled = builder
-                .comment("Turns tree generation on or off")
-                .translation(MoreFeatures.MOD_ID + ".config." + "isTreeGenerationEnabled")
-                .define("isTreeGenerationEnabled", true);
-        builder.pop();
-
         builder.push("Enchantment Settings");
         savingGraceChanceIncrease = builder
                 .comment("The amount of increase in chance every level (e.g 20% for level 1, 40% for level 2)")