+++ src/main/java/io/github/xf8b/morefeatures/items/tools/corrupted/CorruptedPickaxe.java
@@ -23,15 +23,8 @@ import java.util.List;
 import java.util.concurrent.ThreadLocalRandom;
 
 public class CorruptedPickaxe extends PickaxeItem {
-    private static int randomAttackDamage = ThreadLocalRandom.current().nextInt(
-            0, MoreFeaturesConfig.corruptedToolAttackDamageMax + 1
-    );
-    private static float randomAttackSpeed = (float) ThreadLocalRandom.current().nextDouble(
-            0, MoreFeaturesConfig.corruptedToolAttackSpeedMax + 1
-    );
-
     public CorruptedPickaxe() {
-        super(MoreFeaturesItemTier.CORRUPTED, randomAttackDamage, randomAttackSpeed, new Item.Properties()
+        super(MoreFeaturesItemTier.CORRUPTED, 0, 0, new Item.Properties()
                 .maxStackSize(1)
                 .group(MoreFeatures.instance.itemGroup)
         );