+++ src/main/java/io/github/xf8b/morefeatures/world/feature/LemonTree.java
@@ -13,13 +13,14 @@ import javax.annotation.Nullable;
 import java.util.Random;
 
 public class LemonTree extends Tree {
-    public static final TreeFeatureConfig LEMON_TREE_CONFIG = (new TreeFeatureConfig.Builder(
+    public static final TreeFeatureConfig LEMON_TREE_CONFIG = new TreeFeatureConfig.Builder(
             new SimpleBlockStateProvider(MoreFeaturesRegistries.LEMON_LOG.get().getDefaultState()),
             new SimpleBlockStateProvider(MoreFeaturesRegistries.LEMON_LEAVES.get().getDefaultState()),
-            new BlobFoliagePlacer(2, 0)))
+            new BlobFoliagePlacer(2, 0))
             .baseHeight(4)
             .heightRandA(2)
-            .foliageHeight(2)
+            .foliageHeight(3)
+            .ignoreVines()
             .setSapling((IPlantable) MoreFeaturesRegistries.LEMON_SAPLING.get())
             .build();